+++ crates/api/src/measurement/v2/ops.rs
@@ -1,7 +1,12 @@
+use std::cmp::Ordering;
+
 use approx::ulps_eq;
 
 use crate::{
-    v2::{dim::IsCommensurableWith, ops::CommensurableEq},
+    v2::{
+        dim::IsCommensurableWith,
+        ops::{CommensurableEq, CommensurableOrd},
+    },
     Composition, Measurement, UcumUnit,
 };
 
@@ -12,6 +17,16 @@ impl CommensurableEq<Composition> for Measurement {
         }
 
         Some(ulps_eq!(self.scalar(), other.scalar()))
+    }
+}
+
+impl CommensurableOrd<Composition> for Measurement {
+    fn commensurable_cmp(&self, other: &Self) -> Option<Ordering> {
+        if !self.is_commensurable_with(other) {
+            return None;
+        }
+
+        self.scalar().partial_cmp(&other.scalar())
     }
 }
 
@@ -37,5 +52,21 @@ mod tests {
 
         let rhs = Measurement::new(1000.0, GRAM);
         assert_eq!(lhs.commensurable_eq(&rhs), None);
+    }
+
+    #[test]
+    fn commensurable_ord_test() {
+        let lhs = Measurement::new(1000.0, METER);
+        let rhs = Measurement::new(1.0, KILOMETER);
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Equal));
+
+        let rhs = Measurement::new(1.1, KILOMETER);
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Less));
+
+        let rhs = Measurement::new(0.9, KILOMETER);
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Greater));
+
+        let rhs = Measurement::new(1000.0, GRAM);
+        assert_eq!(lhs.commensurable_cmp(&rhs), None);
     }
 }