+++ crates/api/src/term/v2/ops.rs
@@ -1,7 +1,12 @@
+use std::cmp::Ordering;
+
 use approx::ulps_eq;
 
 use crate::{
-    v2::{dim::IsCommensurableWith, ops::CommensurableEq},
+    v2::{
+        dim::IsCommensurableWith,
+        ops::{CommensurableEq, CommensurableOrd},
+    },
     Composition, Term, UcumUnit,
 };
 
@@ -12,6 +17,16 @@ impl CommensurableEq<Composition> for Term {
         }
 
         Some(ulps_eq!(self.scalar(), other.scalar()))
+    }
+}
+
+impl CommensurableOrd<Composition> for Term {
+    fn commensurable_cmp(&self, other: &Self) -> Option<Ordering> {
+        if !self.is_commensurable_with(other) {
+            return None;
+        }
+
+        self.scalar().partial_cmp(&other.scalar())
     }
 }
 
@@ -32,5 +47,22 @@ mod tests {
         let lhs = term!(Meter);
         let rhs = term!(Gram);
         assert!(lhs.commensurable_eq(&rhs).is_none());
+    }
+
+    #[test]
+    fn commensurable_ord_test() {
+        let lhs = term!(Meter, factor: 1000);
+        let rhs = term!(Kilo, Meter);
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Equal));
+
+        let lhs = term!(Meter, factor: 10);
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Less));
+
+        let lhs = term!(Meter, factor: 1_000_000);
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Greater));
+
+        let lhs = term!(Meter);
+        let rhs = term!(Gram);
+        assert_eq!(lhs.commensurable_cmp(&rhs), None);
     }
 }