+++ crates/api/src/unit/v2/ops.rs
@@ -1,7 +1,12 @@
+use std::cmp::Ordering;
+
 use approx::ulps_eq;
 
 use crate::{
-    v2::{dim::IsCommensurableWith, ops::CommensurableEq},
+    v2::{
+        dim::IsCommensurableWith,
+        ops::{CommensurableEq, CommensurableOrd},
+    },
     Composition, UcumUnit, Unit,
 };
 
@@ -12,6 +17,16 @@ impl CommensurableEq<Composition> for Unit {
         }
 
         Some(ulps_eq!(self.scalar(), other.scalar()))
+    }
+}
+
+impl CommensurableOrd<Composition> for Unit {
+    fn commensurable_cmp(&self, other: &Self) -> Option<Ordering> {
+        if !self.is_commensurable_with(other) {
+            return None;
+        }
+
+        self.scalar().partial_cmp(&other.scalar())
     }
 }
 
@@ -32,5 +47,22 @@ mod tests {
         let lhs = unit!(term!(Meter));
         let rhs = unit!(term!(Gram));
         assert!(lhs.commensurable_eq(&rhs).is_none());
+    }
+
+    #[test]
+    fn commensurable_ord_test() {
+        let lhs = unit!(term!(Meter, factor: 1000));
+        let rhs = unit!(term!(Kilo, Meter));
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Equal));
+
+        let lhs = unit!(term!(Meter, factor: 10));
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Less));
+
+        let lhs = unit!(term!(Meter, factor: 1_000_000));
+        assert_eq!(lhs.commensurable_cmp(&rhs), Some(Ordering::Greater));
+
+        let lhs = unit!(term!(Meter));
+        let rhs = unit!(term!(Gram));
+        assert_eq!(lhs.commensurable_cmp(&rhs), None);
     }
 }