+++ crates/api/src/v2/ops.rs
@@ -1,4 +1,4 @@
-use std::ops::Mul;
+use std::{cmp::Ordering, ops::Mul};
 
 use super::dim::IsCommensurableWith;
 
@@ -28,5 +28,57 @@ where
     #[inline]
     fn commensurable_ne(&self, other: &Rhs) -> Option<bool> {
         self.commensurable_eq(other).map(|r| !r)
+    }
+}
+
+/// Similar to `std::cmp::PartialOrd`, but allows for checking if types are commensurable before
+/// ordering. This makes room for types to use `std::cmp::PartialOrd` for strict object comparison
+/// (particularly when considering hash-related operations; we want to have the option to have the
+/// hash for, say, a `Term` that represents `1000m` to be different than a `Term` that represents
+/// `1km`).
+///
+pub trait CommensurableOrd<D, Rhs = Self>: IsCommensurableWith<D, Rhs>
+where
+    Rhs: IsCommensurableWith<D>,
+    D: PartialEq + Mul<i32, Output = D>,
+{
+    /// Similar to  `std::cmp::PartialOrd::partial_cmp()`, this is where you should a) check if
+    /// `self` and `other` are commensurable (using `wise_units::v2::dim::IsCommensurableWith::is_commensurable_with()`),
+    /// then proceed to determining the order of the two values.
+    ///
+    fn commensurable_cmp(&self, other: &Rhs) -> Option<Ordering>;
+
+    /// Tests less than (for `self` and `other`).
+    ///
+    #[inline]
+    fn lt(&self, other: &Rhs) -> bool {
+        matches!(self.commensurable_cmp(other), Some(Ordering::Less))
+    }
+
+    /// Tests less than or equal to (for `self` and `other`).
+    ///
+    #[inline]
+    fn le(&self, other: &Rhs) -> bool {
+        matches!(
+            self.commensurable_cmp(other),
+            Some(Ordering::Less | Ordering::Equal)
+        )
+    }
+
+    /// Tests greater than (for `self` and `other`).
+    ///
+    #[inline]
+    fn gt(&self, other: &Rhs) -> bool {
+        matches!(self.commensurable_cmp(other), Some(Ordering::Greater))
+    }
+
+    /// Tests greater than or equal to (for `self` and `other`).
+    ///
+    #[inline]
+    fn ge(&self, other: &Rhs) -> bool {
+        matches!(
+            self.commensurable_cmp(other),
+            Some(Ordering::Greater | Ordering::Equal)
+        )
     }
 }